% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_stats.R
\name{t_test_stats}
\alias{t_test_stats}
\title{Student's T-test from descriptive statistics.}
\usage{
t_test_stats(
  mean_x,
  std_x,
  n_x,
  mean_y,
  std_y,
  n_y,
  alternative = c("two.sided", "less", "greater"),
  mu = 0,
  var.equal = FALSE,
  conf.level = 0.95
)
}
\arguments{
\item{mean_x}{a mean of sample x.}

\item{std_x}{a corrected sample standard deviation of sample x.}

\item{n_x}{a number of observations of sample x.}

\item{mean_y}{an optional mean of sample y.}

\item{std_y}{an optional corrected sample standard deviation of sample y.}

\item{n_y}{an optional number of observations of sample y.}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of \code{"two.sided"} (default), \code{"greater"} or \code{"less"}. You can specify just the initial letter.}

\item{mu}{a number indicating the true value of the mean (or difference in means if you are performing a two sample test).}

\item{var.equal}{a logical variable indicating whether to treat the two variances as being equal. If TRUE then the pooled variance is used to estimate the variance otherwise the Welch (or Satterthwaite) approximation to the degrees of freedom is used.}

\item{conf.level}{confidence level of the interval.}
}
\description{
Performs one and two sample t-tests from descriptive statistics.
In short, it is like \code{python}'s \code{scipy.stats.ttest_ind_from_stats}.
This is partially based on the code of \code{\link[stats]{t.test}} in \code{stats} on R 4.2.2 by R Core Team.
}
\details{
For a one-sample only test, the mean and standard deviation of one sample and the number of samples are required.
For two-group comparisons, the mean, standard deviation, and number of samples for the two groups are all required.
}
\examples{
t_test_stats(25, 2.5, 10, mu = 23)
t_test_stats(25, 2.5, 10, 23, 3.4, 10)
}
\seealso{
\code{\link[stats]{t.test}}
}
